+++ rl_zoo3/utils.py
@@ -199,7 +199,7 @@ def create_test_env(
     :return:
     """
     # Avoid circular import
-    from utils.exp_manager import ExperimentManager
+    from rl_zoo3.exp_manager import ExperimentManager
 
     # Create the environment and wrap it if necessary
     env_wrapper = get_wrapper_class(hyperparams)