+++ CHANGELOG.md
@@ -1,9 +1,29 @@
-## Release 1.6.1a4 (WIP)
+## Release 1.6.2 (2022-10-02)
 
 ### Breaking Changes
+- RL Zoo is now a python package
+- low pass filter was removed
+
+### New Features
+- RL Zoo cli: `rl_zoo3 train` and `rl_zoo3 enjoy`
+
+### Bug fixes
+
+### Documentation
+
+### Other
+
+## Release 1.6.1 (2022-09-30)
+
+**Progress bar and custom yaml file**
+
+### Breaking Changes
+- Upgraded to Stable-Baselines3 (SB3) >= 1.6.1
+- Upgraded to sb3-contrib >= 1.6.1
 
 ### New Features
 - Added `--yaml-file` argument option for `train.py`to read hyperparameters from custom yaml files (@JohannesUl)
+
 ### Bug fixes
 - Added `custom_object` parameter on record_video.py (@Affonso-Gui)
 - Changed `optimize_memory_usage` to `False` for DQN/QR-DQN on record_video.py (@Affonso-Gui)
@@ -17,6 +37,8 @@
 ### Other
 
 ## Release 1.6.0 (2022-08-05)
+
+**RecurrentPPO (ppo_lstm) and Huggingface integration**
 
 ### Breaking Changes
 - Change default value for number of hyperparameter optimization trials from 10 to 500. (@ernestum)